Chat interface

+++ retrival_augment_git.py
@@ -1,34 +1,16 @@
 from embeddings_dataset_langchain import EmbeddingsDataset
-from transformers import AutoModelForCausalLM, BitsAndBytesConfig, AutoTokenizer, TextIteratorStreamer
-
 from langchain_openai import  OpenAIEmbeddings
 from openai import OpenAI
-from threading import Thread
 from zipfile import ZipFile, BadZipFile
 import PROMPTS
 import MODEL_TYPES
 import CONTEXT_SIZE
 import os
-import torch
 import json
 import git
 import subprocess
-import importlib.util
 import argparse
 import shutil
-
-
-def supports_flash_attention():
-    """Check if a GPU supports FlashAttention."""
-    major, minor = torch.cuda.get_device_capability(0)
-    
-    flash_attention = False if  importlib.util.find_spec('flash_attn') is None else True
-    
-    # Check if the GPU architecture is Ampere (SM 8.x) or newer (SM 9.x)
-    is_sm8x = major == 8 and minor >= 0
-    is_sm9x = major == 9 and minor >= 0
-
-    return (is_sm8x or is_sm9x) and flash_attention
 
 class RetrivalAugment:
     
@@ -336,7 +318,7 @@ class RetrivalAugment:
         # Save the cache list     
         json.dump(self.cached, open(self.cache_repo_list, 'w+'), indent=6)
         
-    def _get_relevant_docs(self, git_repos: list[str], versions: list[str], inputs) -> str:
+    def _get_relevant_docs(self, git_repos: list[str], versions: list[str], inputs: list[str]) -> str:
         
         result_string = "### Most Relevant Documents"
         # Iterate over all repositories
@@ -353,7 +335,7 @@ class RetrivalAugment:
                 if f"{repo_rel_dir}/{repo_rel_name}" in version:
                     _, true_ver = os.path.split(version)    
                     relevant_docs = (self.version_specific_documents[repo][true_ver]).relevant_docs_filename(
-                        inputs, 
+                        inputs[-1]['content'], 
                         k=max(1, CONTEXT_SIZE.GIT_DOCUMENTS//len(versions)), 
                         fetch_k=max(1, CONTEXT_SIZE.GIT_DIVERSE_K//len(versions))
                     )
@@ -382,7 +364,7 @@ class RetrivalAugment:
         return result_string
                            
             
-    def __call__(self, git_repos: list[str] = None, versions= None, inputs = '', shared=None, temperature: float= 0.2, api_key:str = None, model:str = 'gpt-3.5-turbo', system_prompt:str = PROMPTS.SYSTEM_PROMPT):    
+    def __call__(self, git_repos: list[str] = None, versions= None, inputs: list[str] = None, shared=None, temperature: float= 0.7, api_key:str = None, model:str = 'gpt-3.5-turbo', system_prompt:str = PROMPTS.SYSTEM_PROMPT):    
         if len(git_repos) == 0 and len(shared) == 0:
             yield 'I was not given any documentation path to work from.'
             return
@@ -393,55 +375,67 @@ class RetrivalAugment:
             yield "No API Key Provided"
             return
         
+        copied_inputs = inputs.copy()
         
-        version_context = []
         # Get most relevant documents for given repos and versions
-        for repo in git_repos:
+        for i, user_in in enumerate(copied_inputs):
+            if user_in['role'] == 'user':
+                for repo in git_repos:
+                    copied_inputs[i]['docs'] = []
             
-            _, repo_rel_name = os.path.split(repo.removesuffix('.git'))
-            repo_dir = os.path.dirname(repo)
-            _, repo_rel_dir =  os.path.split(repo_dir)
-            
-            for version in versions:
-                if f"{repo_rel_dir}/{repo_rel_name}" in version:
-                    _, true_ver = os.path.split(version)
-                    # Get the relevant documents
-                    version_context += self.version_specific_documents[repo][true_ver](inputs, 
-                                                    k=max(1, CONTEXT_SIZE.GIT_DOCUMENTS//len(versions)), 
-                                                    fetch_k=max(1, CONTEXT_SIZE.GIT_DIVERSE_K//len(versions)))
-         
-        shared_context = []
+                    _, repo_rel_name = os.path.split(repo.removesuffix('.git'))
+                    repo_dir = os.path.dirname(repo)
+                    _, repo_rel_dir =  os.path.split(repo_dir)
+
+                    for version in versions:
+                        if f"{repo_rel_dir}/{repo_rel_name}" in version:
+                            _, true_ver = os.path.split(version)
+                            # Get the relevant documents
+                            copied_inputs[i]['docs'] += self.version_specific_documents[repo][true_ver](
+                                user_in['content'], 
+                                k=max(1, CONTEXT_SIZE.GIT_DOCUMENTS//len(versions)), 
+                                fetch_k=max(1, CONTEXT_SIZE.GIT_DIVERSE_K//len(versions))
+                            )
+                            
         # Get most relevant documents for shared documents
-        for share in shared:
-            # Get the relevant documents
-            shared_context += self.shared_documents[share].querry_documents(f"{'' if (versions==None or len(versions) == 0)  else versions}\n{inputs}", 
-                                                                            k=max(1, CONTEXT_SIZE.SHARED_DOCUMENTS//len(shared)), 
-                                                                            fetch_k=max(1, CONTEXT_SIZE.SHARED_DIVERSE_K//len(shared)))      
+        for i, user_in in enumerate(copied_inputs):
+            if user_in['role'] == 'user':
+                copied_inputs[i]['shared'] = []
+                
+                for share in shared:
+                    # Get the relevant documents
+                    copied_inputs[i]['shared'] += self.shared_documents[share].querry_documents(
+                        f"{'' if (versions==None or len(versions) == 0) else versions}\n{user_in['content']}", 
+                        k=max(1, CONTEXT_SIZE.SHARED_DOCUMENTS//len(shared)), 
+                        fetch_k=max(1, CONTEXT_SIZE.SHARED_DIVERSE_K//len(shared))
+                    )
+        
         messages = [
             {
                 'role' : 'system',
                 'content' : system_prompt
-            },
-            {
-                "role": "user",
-                "content":  PROMPTS.INPUT_PROMPT.format(version=versions, 
-                                                            version_context=version_context, 
-                                                            shared_context=shared_context, 
-                                                            inputs=inputs)
-            },
-        ] 
+            }]
+        for one_input in copied_inputs:
+            role = one_input['role']
+            messages.append({
+                "role": one_input['role'],
+                "content": one_input['content'] if role == 'assistant' else PROMPTS.INPUT_PROMPT.format( 
+                    version_context=one_input['docs'], 
+                    shared_context=one_input['shared'], 
+                    question=one_input['content']
+                )
+            })
         open_api = OpenAI(api_key=api_key, base_url=MODEL_TYPES.LLM_MODELS[model])
         completion = open_api.chat.completions.create(
             model=model,
             messages=messages,
-            max_tokens=1024,
+            max_tokens=2048,
             stream=True,
-            temperature=float(temperature) if (temperature != None and temperature > 0)  else 0.2,
+            temperature=float(temperature) if (temperature != None and temperature > 0)  else 0.7,
         )
         
-        partial_message = ""
+        inputs.append({'role': 'assistant', 'content': ''})
         for chunk in completion:
-            partial_message += chunk.choices[0].delta.content if chunk.choices[0].delta.content != None else ''
-            yield partial_message
-
+            inputs[-1]['content'] += chunk.choices[0].delta.content if chunk.choices[0].delta.content != None else ''
+            yield inputs
